BasicEnemy

+++ Assets/Entity/Entity.cs
@@ -7,7 +7,7 @@ public class Entity
         public string entityName;
         public bool male;
         public Stats entityStats;
-        public float experienceDrop;
+        public int experienceDrop;
         public int nbEnemyKill;
         public int nbDeath;
         public int money;
